Reverted back to https://github.com/andrewphorn/ClassiCube-Client/tree/91b87c76d8850db33d967f7bd2e4d3573abc9347 manually readded commits since then

+++ src/main/java/com/mojang/minecraft/level/generator/noise/PerlinNoise.java
@@ -22,21 +22,21 @@ public class PerlinNoise extends Noise {
 	return unknown1 + unknown0 * (unknown2 - unknown1);
     }
 
-    private short[] noise;
+    private int[] noise;
 
     public PerlinNoise() {
 	this(new Random());
     }
 
     public PerlinNoise(Random random) {
-	noise = new short[512];
+	noise = new int[512];
 
-	for (short count = 0; count < 256; noise[count] = count++) {
+	for (int count = 0; count < 256; noise[count] = count++) {
 	}
 
-	for (short count = 0; count < 256; count++) {
-	    short unknown0 = (short) (random.nextInt(256 - count) + count);
-	    short unknown1 = noise[count];
+	for (int count = 0; count < 256; count++) {
+	    int unknown0 = random.nextInt(256 - count) + count;
+	    int unknown1 = noise[count];
 
 	    noise[count] = noise[unknown0];
 	    noise[unknown0] = unknown1;
