+++ src/main/java/com/mojang/minecraft/sound/AudioInfo.java
@@ -3,7 +3,7 @@ package com.mojang.minecraft.sound;
 public class AudioInfo {
     public float volume = 1.0F;
 
-    public int update(int[] var1, int var2) {
+    public int update(short[] var1, int var2) {
 	return 0;
     }
 }
